gui/wsgi/requirements.txt
flask==2.3.2
google-cloud-logging==3.5.0
gunicorn==20.1.0
gevent-websocket
flask_socketio==5.3.4
flask-sockets